Refactor scraping into one function
Existing functions besides those for cleaning and saving data are now private functions called within a single, public scraping function. This function also determines whether a URL is for a winter transfer window.

The GET timeout is increased to 30 s.

When reading soup to a DataFrame, the merging process now concatenates one list of DataFrames instead of two lists (one for transfers in and one for out).